+++ prefs.py
@@ -48,6 +48,7 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
         "pivot point (instead of manually by pressing 'C') to the surface of whatever object you are looking at while using the operator", default = False)
     pivotDig: bpy.props.FloatProperty(name = "Pivot Dig", description = "When relocating the pivot point, specifies how far the pivot will be moved into the surface you are looking at, based on a percentage of its distance to the 3D View camera" , \
         default = 0.0, min = 0.0, max = 100.0, soft_min = 0.0, soft_max = 100.0, step = 100, precision = 0, subtype = "PERCENTAGE")
+    toggleMode: bpy.props.BoolProperty(name = "Toggle strafe-mode instead of hold-to-strafe.", description = "When checked, strafe-mode will only quit when pressing the key a second time or pressing Escape", default = False)
 
     def draw(self, context: bpy.types.Context):
         prefs: MouseStrafingPreferences = context.preferences.addons["mouse_strafing"].preferences
@@ -104,3 +105,6 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
 
         row = box.row()
         row.prop(self, "showCrosshair")
+
+        row = box.row()
+        row.prop(self, "toggleMode")